requirements.txt
requests
audioop-lts
python-dotenv
discord.py-self